Get all tests passing on IE w/ manual polyfill loading
