--- packages/lit-html/development/test/attribute-order_test.html
@@ -1,29 +1,0 @@
-<html>
-  <body>
-    <script>
-      const originalAttributesGetter = Object.getOwnPropertyDescriptor(Element.prototype, 'attributes');
-      Object.defineProperty(Element.prototype, 'attributes', {
-        get() {
-          const attributes = originalAttributesGetter.get.call(this);
-          const proxy = new Proxy(attributes, {
-            get(target, prop, receiver) {
-              if (typeof prop === 'number' || /^[0-9]+$/.test(prop)) {
-                // reverse the order of attributes
-                return target[target.length - prop - 1];
-              }
-              return target[prop];
-            }
-          });
-          return proxy;
-        }
-      });
-    </script>
-    <script type="module">
-      import { runTests } from '@web/test-runner-mocha';
-
-      runTests(async () => {
-        await import('./lit-html_test.js');
-      });
-    </script>
-  </body>
-</html>
