+++ packages/lit-html/src/test/directives/class-map_test.ts
@@ -15,6 +15,7 @@
 import {html, svg, render} from '../../lit-html.js';
 import {ClassInfo, classMap} from '../../directives/class-map.js';
 import {assert} from '@esm-bundle/chai';
+import '../polyfills.js';
 
 suite('classMap directive', () => {
   let container: HTMLDivElement;
