+++ packages/lit-html/src/test/directives/repeat_test.ts
@@ -12,12 +12,11 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-// import {guard} from '../../directives/guard.js';
 import {repeat} from '../../directives/repeat.js';
 import {render, html} from '../../lit-html.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
-
 import {assert} from '@esm-bundle/chai';
+import '../polyfills.js';
 
 function assertItemIdentity(
   oldChildren: HTMLElement[],
