+++ packages/lit-html/src/test/directives/template-content_test.ts
@@ -16,6 +16,7 @@ import {templateContent} from '../../directives/template-content.js';
 import {html, render} from '../../lit-html.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
 import {assert} from '@esm-bundle/chai';
+import '../polyfills.js';
 
 suite('templateContent', () => {
   let container: HTMLElement;
