+++ packages/lit-html/src/test/directives/unsafe-html_test.ts
@@ -16,6 +16,7 @@ import {unsafeHTML} from '../../directives/unsafe-html.js';
 import {render, html} from '../../lit-html.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
 import {assert} from '@esm-bundle/chai';
+import '../polyfills.js';
 
 suite('unsafeHTML directive', () => {
   let container: HTMLElement;
