+++ packages/lit-html/src/test/parts_test.ts
@@ -27,6 +27,7 @@ import {
   removePart,
   restoreNodePart,
 } from '../parts.js';
+import './polyfills.js';
 
 suite('lit-html', () => {
   let container: HTMLDivElement;
