+++ packages/tests/package.json
@@ -12,8 +12,8 @@
   "devDependencies": {
     "@web/dev-server-legacy": "^0.1.4",
     "@web/dev-server-rollup": "^0.2.7",
-    "@web/test-runner": "^0.7.35",
+    "@web/test-runner": "^0.7.39",
     "@web/test-runner-playwright": "^0.5.7",
-    "@web/test-runner-saucelabs": "~0.0.7"
+    "@web/test-runner-saucelabs": "0.0.9"
   }
 }
