+++ packages/tests/web-test-runner.config.js
@@ -35,7 +35,7 @@ const browserPresets = {
     'sauce:Windows 10/chrome@latest-3',
     'sauce:macOS 10.15/safari@latest',
     // "sauce:Windows 10/MicrosoftEdge@18", // Browser start timeout
-    // "sauce:Windows 7/internet explorer@11", // Browser start timeout
+    "sauce:Windows 7/internet explorer@11", // Browser start timeout
   ],
 };
